% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{chebyshev_distance}
\alias{chebyshev_distance}
\title{Chebyshev Distance}
\usage{
chebyshev_distance(point1, point2)
}
\arguments{
\item{point1}{A numeric vector representing the first point.}

\item{point2}{A numeric vector representing the second point.}
}
\value{
The Chebyshev distance (a scalar value).
}
\description{
Calculate the Chebyshev distance between two points in n-dimensional space.
}
\examples{
chebyshev_distance(c(1, 2, 3), c(4, 5, 6))
}
