% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Cosine Distance between Two Vectors}
\usage{
cosine_distance(point1, point2)
}
\arguments{
\item{point1}{A numeric vector representing the first vector.}

\item{point2}{A numeric vector representing the second vector.}
}
\value{
The cosine distance (a scalar value).
}
\description{
Calculate the cosine distance between two vectors.
}
\examples{
point1 <- c(1, 2, 3)
point2 <- c(4, 5, 6)
cosine_distance(point1, point2)
}
