% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{fishers_score}
\alias{fishers_score}
\title{Compute Fisher Score for Features in filter methods.}
\usage{
fishers_score(data, labels)
}
\arguments{
\item{data}{A data frame containing the independent variables (features).}

\item{labels}{A vector of class labels corresponding to the samples in the data frame.}
}
\value{
A data frame containing Fisher Scores for each feature.
}
\description{
This function calculates the Fisher Score for each feature in a dataset
based on the provided class labels. The Fisher Score is a measure of
the significance of a feature in discriminating between different classes.
}
\details{
The Fisher Score (F_s) is computed using the formula:
\deqn{F_s = \frac{\sum_{j=1}^{K} p_j(\mu_{j}^{i}-\mu^i)^2}{\sum_{j=1}^{K}p_j (\sigma_{j}^i)^{2}}}

where \eqn{\mu_{j}^{i}} and \eqn{\sigma_{j}^i} are the mean and standard deviation of the \eqn{j^-th} -class and \eqn{i^th} feature,
\eqn{p_j}  is the proportion  of data points of class belonging to the class \eqn{j}.
Greater Fisher's score values indicate greater discriminating power of the feature.
}
\examples{
data(iris)
iris_data <- iris[, 1:4]
iris_labels <- iris$Species
fisher_score_df <- fishers_score(iris_data, iris_labels)
print(fisher_score_df)
}
