% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{gini_index_score}
\alias{gini_index_score}
\title{gini index}
\usage{
gini_index_score(probabilities)
}
\arguments{
\item{probabilities}{Gini index, a measure of impurity or inequality, for a set of probabilities.
\deqn{G(v_i) = 1 - \sum_{j=1}^{k} p_j^2}.}
}
\description{
gini index
}
