% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hopkins_stat}
\alias{hopkins_stat}
\title{Hopkins Statistics}
\usage{
hopkins_stat(
  X,
  m = ceiling(nrow(X)/10),
  d = ncol(X),
  k = 1,
  U = NULL,
  method = "simple"
)
}
\value{
Data (matrix or data.frame) to check clusterability.
}
\description{
\deqn{H = \frac{\sum_{i=1}^{r} \beta_i}{\sum_{i=1}^{r}(\alpha_i+\beta_i) }}
}
\examples{
set.seed(1)
hopkins(iris[, 1:4], m=15) # .9952293

hop <- rep(NA, 100)
for(i in 1:100){
  hop[i] <- hopkins(iris[,1:4], m=8)
}
mean(hop)
}
