% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Manhattan Distance}
\usage{
manhattan_distance(point1, point2)
}
\arguments{
\item{point1}{numeric vector representing the first point.}

\item{point2}{numeric vector representing the second point.}
}
\value{
The Manhattan distance (a scalar value).
}
\description{
Calculate the Manhattan distance between two points in n-dimensional space.
}
\details{
The Manhattan distance distance between two points
calculated as:\deqn{S(x,y) = \sum_{i=1}^{n} |x_i - y_i|}
}
\examples{
manhattan_distance(c(1, 2, 3), c(4, 5, 6))
}
