% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{minkowski_distance}
\alias{minkowski_distance}
\title{Minkowski Distance}
\usage{
minkowski_distance(vector1, vector2, p_value)
}
\arguments{
\item{vector1}{A numeric vector representing the first point.}

\item{vector2}{A numeric vector representing the second point.}

\item{p_value}{A numeric value representing the order of the Minkowski distance.}
}
\value{
The Minkowski distance between \code{vector1} and \code{vector2} using the
specified \code{p_value}.
}
\description{
This function calculates the Minkowski distance between two numeric vectors
\code{vector1} and \code{vector2}, using the parameter \code{p_value} to determine
the order of the distance.
}
\details{
calculated as
\deqn{S(x,y) =  (\sum\limits_{i=1}^{n} |x_i - y_i|^p)^{1/p}}
}
\examples{
# Example usage
vector1 <- c(1, 2, 3)
vector2 <- c(4, 5, 6)
p_value <- 2
distance <- minkowski_distance(vector1, vector2, p_value)
}
