% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stirling_second}
\alias{stirling_second}
\title{Calculate Stirling Numbers of the Second Kind}
\usage{
stirling_second(n, k)
}
\arguments{
\item{n}{The total number of objects.}

\item{k}{The number of non-empty subsets.}
}
\value{
The Stirling number of the second kind \eqn{S(n, k)}.
}
\description{
A brute-force or exhaustive algorithm for finding a good clustering is to simply
generate all possible partitions of \eqn{n} points into \eqn{k} clusters, evaluate some optimization
score for each of them, and to retain the clustering that yields the best score. The
exact number of ways of partitioning \eqn{n} points into \eqn{k} non-empty and disjoint parts
is given by the Stirling numbers of the second kind,
}
\details{
\deqn{S(n,k) = \frac{1}{k!}\sum_{t=0}^{k} (-1)^t \binom{a}{b} (k-t)^n}

Each point can be assigned to any one of the \eqn{k} clusters, so there are at
most \eqn{k^n} possible clusterings. However, any permutation of the \eqn{n} clusters within a
given clustering yields an equivalent clustering, therefore, there are \eqn{O(k^n/k!)}  clusterings of n points into \eqn{k} groups
}
\examples{
stirling_second(5, 3)

}
\references{
Wikipedia contributors. (2021, June 12). Stirling numbers of the second kind. In Wikipedia, The Free Encyclopedia.
Retrieved from \url{https://en.wikipedia.org/wiki/Stirling_numbers_of_the_second_kind}
}
